Temporary fix to judge voting and added CRUD files for a judge collection
Issue: If one judge voted and then another judge voted later, the second judge would get an unknown document error. The error is because the second judge has an outdated instance of the hackathon. The temporary fix is that before submitting a vote, the instance of the hackathon is updated on the judge's end before sending a PUT request that the judge has submitted a new vote. However, this does not fix the issue if two judges were to vote at approximately the same time.

Currently creating a judge collection which would handle the one-vote verification (currently the hackathon collection handles verification).