New Judge collection
Created new judge collection which stores judges (containing email, id, and their votes). App will look in this collection to see if a judge has already voted for a project and update when the judge votes for a project. This will fix the issue if two judges vote at approximately the same time, there won't be "Document not found" error. ID verification is still done in the hackathon collection.

Also upon sending emails, the app will generate judges in the Judge collection and still append the judges to a hackathon (like before).