Comments, confirmations, and overwriting judges
Added comments to enhance code visibility. Added confirmations for actions that would be irreversible and damaging if unintentional.

If admin uploads and wants to send emails to an updated CSV file of judges, the old judges will be deleted from the hackathon collection and the judge collection. Also there is a confirmation if the admin wants to send new emails.